refactor(rust): fix message types, make transport use one socket for all comms
- Message: now support the following types: Udp, Channel, Worker
- Message: has a new field, message_type, to indicate what is in the message
           body. The following types are recognized: Ping, Pong, Payload,
           KeyAgreementM1..M3
- UdpTransport: now performs all io over the same socket. When creating a
                new transport, provide the local address/port to bind to.
- UdpTransport: The "Add" command is no longer supported. The transport will
                attempt to send the message to whatever address is in the
                route.
