+++ implementations/rust/node/Cargo.toml
@@ -17,3 +17,4 @@ ockam-router = { version = "0.1", path = "../router" }
 ockam-transport = { version = "0.1", path = "../transport"}
 ockam-message = { version = "0.1", path = "../message"}
 ockam-common = { version = "0.1", path = "../common"}
+hex = "0.4.2"
